pyproject.toml
[project]
name = "ableton-record-hue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-hue-v2>=2.2.1",
    "python-osc>=1.9.3",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
